Split proposed examples and requirement tests into their own files.
I wanted a bit more focus when it came to writing test requirements
and proposing possible conventions.

I also wanted to see if I could check coverage... Firstly with NYC
but if not then perhaps by writing some custom functions to check
if all values in the test requirements exist in the proposal (and
highlight any gaps) and to check if all values in the proposal
are being tested. Two-way coverage check.